refact: extract common method into utility class
